P2P add PeerServer class
